read apple calendar and store as unique entity (#7)
The basics for syncing EK events to sqlite. It's a starting point but nowhere near a final thing. The EK handling needs a change to make it testable (i.e. remove the main loop and only start it when needed). Also I need to research libraries that make Ports handling easier and maybe introduces a shim. Like Porcelain, Rambo or MuonTrap. It's fine for now to continue though